+++ config/index.js
@@ -2,6 +2,7 @@ const invoiceNetwork = require('./invoice_network');
 const isActivityEnabled = require('./is_activity_enabled');
 const isConnectEnabled = require('./is_connect_enabled');
 const isEmailConfigured = require('./is_email_configured');
+const isInvoiceEnabled = require('./is_invoice_enabled');
 const isRelayConfigured = require('./is_relay_configured');
 const isSmsConfigured = require('./is_sms_configured');
 const validateServerConfig = require('./validate_server_config');
@@ -11,6 +12,7 @@ module.exports = {
   isActivityEnabled,
   isConnectEnabled,
   isEmailConfigured,
+  isInvoiceEnabled,
   isRelayConfigured,
   isSmsConfigured,
   validateServerConfig,